+++ src/main/java/com/github/storytime/service/PbInternalTransferInfoService.java
@@ -23,7 +23,7 @@ public class PbInternalTransferInfoService {
         this.regExpService = regExpService;
     }
 
-    public String generateIdForFromTransfer(final AppUser u, final Statement cardNum, final Float opAmount, final String comment) {
+    public String generateIdForFromTransfer(final AppUser u, final Statement cardNum, final Double opAmount, final String comment) {
         final String card = String.valueOf(cardNum.getCard());
         return u.getId() +
                 left(card, CARD_TWO_DIGITS) +
@@ -35,7 +35,7 @@ public class PbInternalTransferInfoService {
 
     }
 
-    public String generateIdForToTransfer(final AppUser u, final Statement cardNum, final Float opAmount, final String comment) {
+    public String generateIdForToTransfer(final AppUser u, final Statement cardNum, final Double opAmount, final String comment) {
         final String card = String.valueOf(cardNum.getCard());
         return u.getId() +
                 regExpService.getCardFirstDigits(comment) +
